+++ app/src/main/kotlin/org/breezyweather/sources/RefreshHelper.kt
@@ -974,7 +974,7 @@ class RefreshHelper @Inject constructor(
                         },
                     currentHour,
                     currentDay,
-                    weatherWrapperCompleted.airQuality?.current
+                    weatherWrapperCompleted.airQuality?.current?.toValid()
                         ?: if (isUpdateStillValid(base.currentUpdateTime, wait = 30)) {
                             // Allow to re-use current data if it was successfully refreshed less than 30 min ago
                             location.weather?.current?.airQuality