+++ weather-unit/src/main/kotlin/org/breezyweather/unit/pollen/PollenConcentration.kt
@@ -133,6 +133,13 @@ value class PollenConcentration internal constructor(
         require(decimals >= 0) { "decimals must be not negative, but was $decimals" }
         return formatToExactDecimals(toDouble(unit), decimals.coerceAtMost(unit.decimals.long)) + unit.id
     }
+
+    /**
+     * Return null if the value is not positive, otherwise this value
+     */
+    fun toValidOrNull(): PollenConcentration? {
+        return takeIf { rawValue >= 0 }
+    }
 }
 
 // constructing from number of units